timer

+++ src/config/timer.rs
@@ -5,13 +5,13 @@ pub struct TimerConfig {
     #[serde(default)]
     show_millis: bool,
     #[serde(default)]
-    work_duration: u16,
+    work_duration: u32,
     #[serde(default)]
-    break_duration: u16,
+    break_duration: u32,
     #[serde(default)]
-    long_break_duration: u16,
+    long_break_duration: u32,
     #[serde(default)]
-    long_break_interval: u16,
+    long_break_interval: u32,
 }
 
 impl Default for TimerConfig {
@@ -31,23 +31,23 @@ impl TimerConfig {
         self.show_millis
     }
 
-    pub fn work_duration(&self) -> u64 {
-        self.work_duration as u64 * 60 * 1000
+    pub fn work_duration(&self) -> u32 {
+        self.work_duration as u32 * 60 * 1000
     }
 
-    pub fn break_duration(&self) -> u64 {
-        self.break_duration as u64 * 60 * 1000
+    pub fn break_duration(&self) -> u32 {
+        self.break_duration as u32 * 60 * 1000
     }
 
-    pub fn long_break_duration(&self) -> u64 {
-        self.long_break_duration as u64 * 60 * 1000
+    pub fn long_break_duration(&self) -> u32 {
+        self.long_break_duration as u32 * 60 * 1000
     }
 
-    pub fn long_break_interval(&self) -> u16 {
+    pub fn long_break_interval(&self) -> u32 {
         self.long_break_interval
     }
 
-    pub fn tick_interval(&self) -> u64 {
+    pub fn tick_interval(&self) -> u32 {
         if self.show_millis { 10 } else { 1000 }
     }
 }
